Restructuring project and review comments
