+++ internal/collection/entry/metadata.go
